Add database controller
